+++ setup.cfg
@@ -1,0 +1,16 @@
+[metadata]
+name = pysearcherror
+version = 0.0.1
+author = MPADlein, plantcraft94
+description = A package to search error in python code
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/mpadlein/maybe-it-useful
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+python_requires = >=3.6
+include_package_data = True
